+++ src/core/model/object.cc
@@ -59,16 +59,27 @@ bool
 Object::AggregateIterator::HasNext() const
 {
     NS_LOG_FUNCTION(this);
-    return m_current < m_object->m_aggregates->n;
+    return (m_current < m_object->m_aggregates->n) ||
+           (m_uniAggrIter != m_object->m_unidirectionalAggregates.end());
 }
 
 Ptr<const Object>
 Object::AggregateIterator::Next()
 {
     NS_LOG_FUNCTION(this);
-    Object* object = m_object->m_aggregates->buffer[m_current];
-    m_current++;
-    return object;
+    if (m_current < m_object->m_aggregates->n)
+    {
+        Object* object = m_object->m_aggregates->buffer[m_current];
+        m_current++;
+        return object;
+    }
+    else if (m_uniAggrIter != m_object->m_unidirectionalAggregates.end())
+    {
+        auto object = *m_uniAggrIter;
+        m_uniAggrIter++;
+        return object;
+    }
+    return nullptr;
 }
 
 Object::AggregateIterator::AggregateIterator(Ptr<const Object> object)
@@ -76,6 +87,7 @@ Object::AggregateIterator::AggregateIterator(Ptr<const Object> object)
       m_current(0)
 {
     NS_LOG_FUNCTION(this << object);
+    m_uniAggrIter = object->m_unidirectionalAggregates.begin();
 }
 
 TypeId
@@ -127,6 +139,7 @@ Object::~Object()
         std::free(m_aggregates);
     }
     m_aggregates = nullptr;
+    m_unidirectionalAggregates.clear();
 }
 
 Object::Object(const Object& o)
@@ -153,6 +166,7 @@ Object::DoGetObject(TypeId tid) const
     NS_LOG_FUNCTION(this << tid);
     NS_ASSERT(CheckLoose());
 
+    // First check if the object is in the normal aggregates.
     uint32_t n = m_aggregates->n;
     TypeId objectTid = Object::GetTypeId();
     for (uint32_t i = 0; i < n; i++)
@@ -177,6 +191,20 @@ Object::DoGetObject(TypeId tid) const
             UpdateSortedArray(m_aggregates, i);
             // finally, return the match
             return const_cast<Object*>(current);
+        }
+    }
+
+    // Next check if it's a unidirectional aggregate
+    for (auto& uniItem : m_unidirectionalAggregates)
+    {
+        TypeId cur = uniItem->GetInstanceTypeId();
+        while (cur != tid && cur != objectTid)
+        {
+            cur = cur.GetParent();
+        }
+        if (cur == tid)
+        {
+            return uniItem;
         }
     }
     return nullptr;
@@ -204,6 +232,17 @@ restart:
             current->DoInitialize();
             current->m_initialized = true;
             goto restart;
+        }
+    }
+
+    // note: no need to restart because unidirectionally aggregated objects
+    // can not change the status of the actual object.
+    for (auto& uniItem : m_unidirectionalAggregates)
+    {
+        if (!uniItem->m_initialized)
+        {
+            uniItem->DoInitialize();
+            uniItem->m_initialized = true;
         }
     }
 }
@@ -237,6 +276,17 @@ restart:
             current->DoDispose();
             current->m_disposed = true;
             goto restart;
+        }
+    }
+
+    // note: no need to restart because unidirectionally aggregated objects
+    // can not change the status of the actual object.
+    for (auto& uniItem : m_unidirectionalAggregates)
+    {
+        if (!uniItem->m_disposed && uniItem->GetReferenceCount() == 1)
+        {
+            uniItem->DoDispose();
+            uniItem->m_disposed = true;
         }
     }
 }
@@ -280,11 +330,13 @@ Object::AggregateObject(Ptr<Object> o)
     {
         aggregates->buffer[m_aggregates->n + i] = other->m_aggregates->buffer[i];
         const TypeId typeId = other->m_aggregates->buffer[i]->GetInstanceTypeId();
+        // note: DoGetObject scans also the unidirectional aggregates
         if (DoGetObject(typeId))
         {
             NS_FATAL_ERROR("Object::AggregateObject(): "
                            "Multiple aggregation of objects of type "
-                           << other->GetInstanceTypeId() << " on objects of type " << typeId);
+                           << other->GetInstanceTypeId() << " on objects of type "
+                           << GetInstanceTypeId());
         }
         UpdateSortedArray(aggregates, m_aggregates->n + i);
     }
@@ -323,7 +375,47 @@ Object::AggregateObject(Ptr<Object> o)
     std::free(b);
 }
 
-/**
+void
+Object::UnidirectionalAggregateObject(Ptr<Object> o)
+{
+    NS_LOG_FUNCTION(this << o);
+    NS_ASSERT(!m_disposed);
+    NS_ASSERT(!o->m_disposed);
+    NS_ASSERT(CheckLoose());
+    NS_ASSERT(o->CheckLoose());
+
+    Object* other = PeekPointer(o);
+
+    const TypeId typeId = other->GetInstanceTypeId();
+    // note: DoGetObject scans also the unidirectional aggregates
+    if (DoGetObject(typeId))
+    {
+        NS_FATAL_ERROR("Object::UnidirectionalAggregateObject(): "
+                       "Multiple aggregation of objects of type "
+                       << other->GetInstanceTypeId() << " on objects of type "
+                       << GetInstanceTypeId());
+    }
+
+    m_unidirectionalAggregates.emplace_back(other);
+
+    // Finally, call NotifyNewAggregate on all the objects aggregates by this object.
+    // We skip the aggregated Object and its aggregates because they are not
+    // mutually aggregated to the others.
+    // Unfortunately, we have to make a copy of the aggregated objects, because
+    // NotifyNewAggregate might change it...
+
+    std::list<Object*> aggregates;
+    for (uint32_t i = 0; i < m_aggregates->n; i++)
+    {
+        aggregates.emplace_back(m_aggregates->buffer[i]);
+    }
+    for (auto& item : aggregates)
+    {
+        item->NotifyNewAggregate();
+    }
+}
+
+/*
  * This function must be implemented in the stack that needs to notify
  * other stacks connected to the node of their presence in the node.
  */